moved PeriodSelector to general class file
