+++ composeApp/src/androidMain/kotlin/me/knighthat/component/screen/HomeSongs.kt
@@ -1,18 +1,26 @@
 package me.knighthat.component.screen
 
 import androidx.compose.runtime.Composable
+import androidx.compose.runtime.MutableState
 import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.saveable.rememberSaveable
 import androidx.compose.runtime.setValue
-import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import it.fast4x.rimusic.R
+import it.fast4x.rimusic.enums.MenuStyle
+import it.fast4x.rimusic.enums.TopPlaylistPeriod
+import it.fast4x.rimusic.ui.components.LocalMenuState
+import it.fast4x.rimusic.ui.components.MenuState
+import it.fast4x.rimusic.ui.components.themed.PeriodMenu
 import it.fast4x.rimusic.utils.autoShuffleKey
+import it.fast4x.rimusic.utils.menuStyleKey
 import it.fast4x.rimusic.utils.rememberPreference
+import it.fast4x.rimusic.utils.topPlaylistPeriodKey
 import me.knighthat.component.tab.toolbar.Descriptive
 import me.knighthat.component.tab.toolbar.DualIcon
 import me.knighthat.component.tab.toolbar.DynamicColor
+import me.knighthat.component.tab.toolbar.Menu
 import me.knighthat.component.tab.toolbar.MenuIcon
 
 @Composable
@@ -41,4 +49,49 @@ fun randomSort(): MenuIcon = object: MenuIcon, DynamicColor, Descriptive {
         get() = stringResource( messageId )
 
     override fun onShortClick() { isFirstColor = !isFirstColor }
+}
+
+class PeriodSelector private constructor(
+    private val periodState: MutableState<TopPlaylistPeriod>,
+    override val menuState: MenuState,
+    override val styleState: MutableState<MenuStyle>
+):  MenuIcon, Descriptive, Menu {
+
+    companion object {
+        @JvmStatic
+        @Composable
+        fun init() = PeriodSelector(
+            rememberPreference( topPlaylistPeriodKey, TopPlaylistPeriod.PastWeek ),
+            LocalMenuState.current,
+            rememberPreference( menuStyleKey, MenuStyle.List )
+        )
+    }
+
+    var period: TopPlaylistPeriod = periodState.value
+        set(value) {
+            periodState.value = value
+            field = value
+        }
+
+    override val iconId: Int = period.iconId
+    override val messageId: Int = R.string.statistics
+    override val menuIconTitle: String
+        @Composable
+        get() = stringResource( messageId )
+
+    fun onDismiss( period: TopPlaylistPeriod ) {
+        this.period = period
+        menuState.hide()
+    }
+
+    @Composable
+    override fun ListMenu() { /* Does nothing */ }
+
+    @Composable
+    override fun GridMenu() { /* Does nothing */ }
+
+    @Composable
+    override fun MenuComponent() = PeriodMenu(::onDismiss)
+
+    override fun onShortClick() = super.onShortClick()
 }
